requirements.txt
Bootstrap_Flask==2.2.0
Flask==2.2.5
Flask_WTF==1.1.1
wtforms[email]==2.3.1
python-dotenv~=1.0.0